Refactor: split auth service file
